refactor(controller): isolate selection orchestration
